Rename TaskTodo to ToDoPage and clean up unused components.
Updated the TaskTodo component to ToDoPage for better clarity and consistency. Removed obsolete components `tasks.jsx` and `todo.jsx`. Adjusted imports and routing to reflect the name change.
