+++ frontend/src/App.jsx
@@ -3,7 +3,7 @@ import {  Routes, Route, Navigate } from "react-router-dom";
 
 import LogIn from "./scenes/login";
 import TodaysAgenda from "./scenes/todaysAgenda";
-import TaskTodo from "./components/task/taskTodo";
+import ToDoPage from "./components/task/ToDoPage";
 import AuthProvider, {PrivateRoute} from "./hooks/AuthContext";
 import VerifyEmail from "./scenes/VerifyEmail";
 import ResetPassword from "./scenes/ResetPassword";
@@ -21,7 +21,7 @@ function App() {
               <Route path="/register" element={<Register/>} />
               <Route element={<PrivateRoute />}>
                   <Route path="/today" element={<TodaysAgenda/>} />
-                  <Route path="/taskTodoPage" element={<TaskTodo/>} />
+                  <Route path="/taskTodoPage" element={<ToDoPage/>} />
                   <Route path="/account-settings" element={<AccountSettings/>} />
                   {/* Tillägg av andra länkar senare när sidorna har byggts på*/}
               </Route>
