--- frontend/src/components/tasks.jsx
@@ -1,1 +1,0 @@
-{/*Komponent för tasks*/}
