--- frontend/src/components/todo.jsx
@@ -1,1 +1,0 @@
-{/*Komponent för ToDo's*/}
